Specular light calculation added.

+++ src/Material.hs
@@ -2,8 +2,10 @@ module Material where
 
 import Color
 
-data Material = DiffusiveMaterial (Color Float) | DiffusiveAndSpecularMaterial (Color Float) Int
+data Material = DiffusiveMaterial (Color Float) | DiffusiveAndSpecularMaterial (Color Float) Float
 
 materialColor :: Material -> Color Float
 materialColor (DiffusiveMaterial c) = c
 materialColor (DiffusiveAndSpecularMaterial c _) = c
+
+
